Split backend into two files separating cell helper methods and base simulation helper functions
