+++ Model/simulation.py
@@ -1,15 +1,9 @@
-import numpy as np
 import random as r
-import math
-import time
-import igraph
-import os
 import csv
 import cv2
 import pickle
-import shutil
+import math
 import psutil
-from numba import cuda
 
 from backend import *
